Moved all icons into a single icons file
